fix: dependency missing
